:sparkles: Play local videos
After downloading videos, a small icon will be visible indicating that
the video is available locally. If the video is available locally,
playing it will result in playing the local file instead of playing the
web-version.

Note that this will currently only work with `youtube-dl` and most of
its derivatives. Also note that restarting the application will reset
the association with the local video, the downloaded video will have to
be deleted and redownloaded again.

Closes #72.